Automatize benchmark execution

+++ src/bitmap.h
@@ -13,7 +13,7 @@ class BitMap
 {
     // ATTRIBUTES
     private:
-        using word_t = uint64_t; //!< Type used for bit vector
+        using word_t = uint32_t; //!< Type used for bit vector
         const size_t word_s = sizeof(word_t)*8; //!< Size in bits of word_t
 
         bitIdx_t _size = 0; //!< Size in bits of the bitmap
